@@ -1,42 +1,51 @@
 package hu.university.etelprojekt.etelprojekt.entity;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import java.io.Serializable;
 
 @Entity
-@Table(name = "address")
-public class Address {
+@Table(name = "addresses") // Optional: specify the table name in the database
+public class Address implements Serializable {
+
+    private static final long serialVersionUID = 1L;
 
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "address_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY) // Automatically generate the addressId (primary key)
     private Long addressId;
 
-    @Column(name = "postal_code", nullable = false)
     private String postalCode;
-
-    @Column(name = "city", nullable = false)
     private String city;
-
-    @Column(name = "street_name", nullable = false)
     private String streetName;
-
-    @Column(name = "public_place_name")
     private String publicPlaceName;
-
-    @Column(name = "house_number")
     private String houseNumber;
-
-    @Column(name = "building")
     private String building;
-
-    @Column(name = "staircase")
     private String staircase;
-
-    @Column(name = "floor")
     private String floor;
-
-    @Column(name = "door")
     private String door;
+
+    // Default constructor
+    public Address() {
+    }
+
+    // Constructor with all fields
+    public Address(Long addressId, String postalCode, String city, String streetName,
+            String publicPlaceName, String houseNumber, String building,
+            String staircase, String floor, String door) {
+        this.addressId = addressId;
+        this.postalCode = postalCode;
+        this.city = city;
+        this.streetName = streetName;
+        this.publicPlaceName = publicPlaceName;
+        this.houseNumber = houseNumber;
+        this.building = building;
+        this.staircase = staircase;
+        this.floor = floor;
+        this.door = door;
+    }
 
     // Getters and Setters
     public Long getAddressId() {
@@ -117,5 +126,13 @@ public class Address {
 
     public void setDoor(String door) {
         this.door = door;
+    }
+
+    @Override
+    public String toString() {
+        return "Address [addressId=" + addressId + ", postalCode=" + postalCode + ", city=" + city +
+                ", streetName=" + streetName + ", publicPlaceName=" + publicPlaceName + ", houseNumber=" +
+                houseNumber + ", building=" + building + ", staircase=" + staircase + ", floor=" + floor +
+                ", door=" + door + "]";
     }
 }
