@@ -1,129 +1,72 @@
 package hu.university.etelprojekt.etelprojekt.entity;
 
-import jakarta.persistence.*;
-
-import java.time.LocalDate;
-import java.util.List;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import java.io.Serializable;
 
 @Entity
-@Table(name = "users")
-public class User {
+@Table(name = "users") // Specify the table name for the User entity
+public class User implements Serializable {
+
+    private static final long serialVersionUID = 1L;
 
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id")
-    private Long userId;
+    private Long id;
 
-    @Column(name = "first_name", nullable = false)
-    private String firstName;
+    private String username;
 
-    @Column(name = "last_name", nullable = false)
-    private String lastName;
-
-    @Column(name = "status")
-    private String status;
-
-    @Column(name = "registration_date", nullable = false)
-    private LocalDate registrationDate;
-
-    @Column(name = "user_type", nullable = false)
-    private UserType userType;
-
-    @Column(name = "phone_number")
-    private String phoneNumber;
-
-    @Column(name = "password", nullable = false)
-    private String password;
-
-    @Column(name = "email", nullable = false, unique = true)
+    private String userType; // Store the userType as a String directly
     private String email;
 
-    @ManyToOne
-    @JoinColumn(name = "address_id")
-    private Address address;
-
-    @OneToMany(mappedBy = "address")
-    private List<User> users;
-
-    // Getters and Setters
-    public Long getUserId() {
-        return userId;
+    // Default constructor
+    public User() {
     }
 
-    public void setUserId(Long userId) {
-        this.userId = userId;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getStatus() {
-        return status;
-    }
-
-    public void setStatus(String status) {
-        this.status = status;
-    }
-
-    public LocalDate getRegistrationDate() {
-        return registrationDate;
-    }
-
-    public void setRegistrationDate(LocalDate registrationDate) {
-        this.registrationDate = registrationDate;
-    }
-
-    public UserType getUserType() {
-        return userType;
-    }
-
-    public void setUserType(UserType userType) {
+    // Constructor with all fields
+    public User(Long user_id, String username, String email, String userType) {
+        this.id = user_id;
+        this.username = username;
         this.userType = userType;
-    }
-
-    public String getPhoneNumber() {
-        return phoneNumber;
-    }
-
-    public void setPhoneNumber(String phoneNumber) {
-        this.phoneNumber = phoneNumber;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
         this.email = email;
     }
 
-    public Address getAddress() {
-        return address;
+    // Getters and Setters
+    public Long getUserId() {
+        return id;
     }
 
-    public void setAddress(Address address) {
-        this.address = address;
+    public void setUserId(Long user_id) {
+        this.id = user_id;
     }
 
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getEmail() {
+        return email; // Getter for email
+    }
+
+    public void setEmail(String email) {
+        this.email = email; // Setter for email
+    }
+
+    public String getUserType() {
+        return userType;
+    }
+
+    public void setUserType(String userType) {
+        this.userType = userType;
+    }
+
+    @Override
+    public String toString() {
+        return "User [userId=" + id + ", username=" + username + ", email=" + email + ", userType=" + userType
+                + "]";
+    }
 }
