@@ -13,30 +13,31 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface AddressRepository extends JpaRepository<Address, Long> {
     // Find an address by its ID (Read)
-    Optional<Address> findById(Long addressId);
+    @SuppressWarnings("null")
+    Optional<Address> findById(Long address_id);
 
     // Find an address by postal code (optional)
-    Optional<Address> findByPostalCode(String postalCode);
+    Optional<Address> findByPostalCode(String postal_code);
 
     // Create or update an address (Save/Update)
     // This is automatically handled by the save() method in JpaRepository
 
     // Delete an address by ID (Delete)
-    void deleteById(Long addressId);
+    void deleteById(@SuppressWarnings("null") Long address_id);
 
     // Update address fields like postal code, city, street name, etc.
     @Modifying
-    @Query("UPDATE Address a SET a.postalCode = :postalCode, a.city = :city, a.streetName = :streetName, a.houseNumber = :houseNumber, a.building = :building, a.staircase = :staircase, a.floor = :floor, a.door = :door WHERE a.addressId = :addressId")
-    void updateAddress(@Param("addressId") Long addressId,
-            @Param("postalCode") String postalCode,
+    @Query("UPDATE Address a SET a.postal_code = :postal_code, a.city = :city, a.street_name = :street_name, a.house_number = :house_number, a.building = :building, a.staircase = :staircase, a.floor = :floor, a.door = :door WHERE a.address_id = :address_id")
+    void updateAddress(@Param("address_id") Long address_id,
+            @Param("postal_code") String postal_code,
             @Param("city") String city,
-            @Param("streetName") String streetName,
-            @Param("houseNumber") String houseNumber,
+            @Param("street_name") String street_name,
+            @Param("house_number") String house_number,
             @Param("building") String building,
             @Param("staircase") String staircase,
             @Param("floor") String floor,
             @Param("door") String door);
 
     // Custom method to check if an address exists by postal code (optional)
-    boolean existsByPostalCode(String postalCode);
+    boolean existsByPostalCode(String postal_code);
 }
