@@ -1,7 +1,7 @@
 package hu.university.etelprojekt.etelprojekt.repository;
 
 import hu.university.etelprojekt.etelprojekt.entity.Allergen;
-import hu.university.etelprojekt.etelprojekt.entity.AllergenType;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,19 +11,20 @@ import java.util.Optional;
 public interface AllergenRepository extends JpaRepository<Allergen, Long> {
 
     // Find allergen by ID
-    Optional<Allergen> findById(Long allergenId);
+    @SuppressWarnings("null")
+    Optional<Allergen> findById(Long allergen_id);
 
     // Find allergen by allergen name (enum value)
-    Optional<Allergen> findByAllergenName(AllergenType allergenName);
+    Optional<Allergen> findByAllergenName(String allergen_name);
 
     // Save or update allergen (handled by JpaRepository)
-    @SuppressWarnings("unchecked")
+    @SuppressWarnings({ "unchecked", "null" })
     Allergen save(Allergen allergen);
 
     // Delete allergen by ID
-    void deleteById(Long allergenId);
+    void deleteById(@SuppressWarnings("null") Long allergen_id);
 
     // Check if allergen exists by name
-    boolean existsByAllergenName(AllergenType allergenName);
+    boolean existsByAllergenName(String allergen_name);
 
 }
