@@ -1,6 +1,7 @@
 package hu.university.etelprojekt.etelprojekt.repository;
 
 import hu.university.etelprojekt.etelprojekt.entity.City;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -13,23 +14,24 @@ import java.util.Optional;
 public interface CityRepository extends JpaRepository<City, Long> {
 
     // Find a city by its ID
-    Optional<City> findById(Long cityId);
+    @SuppressWarnings("null")
+    Optional<City> findById(Long city_id);
 
     // Find all cities by category ID (example: get all cities for a particular
     // category)
-    List<City> findByCategoryId(Long categoryId);
+    List<City> findByCategoryId(Long category_id);
 
     // Find a city by its name (example: used for searching a city)
-    Optional<City> findByCityName(String cityName);
+    Optional<City> findByCityName(String city_name);
 
     // Custom query: Get all cities and their associated categories
-    @Query("SELECT c FROM City c JOIN FETCH c.category WHERE c.cityId = :cityId")
-    Optional<City> findCityWithCategory(@Param("cityId") Long cityId);
+    @Query("SELECT c FROM City c JOIN FETCH c.category WHERE c.city_id = :city_id")
+    Optional<City> findCityWithCategory(@Param("city_id") Long city_id);
 
     // Custom query to delete a city by ID
-    void deleteById(Long cityId);
+    void deleteById(@SuppressWarnings("null") Long city_id);
 
     // Custom query to get all cities by category ID
-    @Query("SELECT c FROM City c WHERE c.category.categoryId = :categoryId")
-    List<City> findCitiesByCategory(@Param("categoryId") Long categoryId);
+    @Query("SELECT c FROM City c WHERE c.category.category_id = :category_id")
+    List<City> findCitiesByCategory(@Param("category_id") Long category_id);
 }
