@@ -1,6 +1,7 @@
 package hu.university.etelprojekt.etelprojekt.repository;
 
 import hu.university.etelprojekt.etelprojekt.entity.Menu;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,6 +12,7 @@ import java.util.Optional;
 public interface MenuRepository extends JpaRepository<Menu, Long> {
 
     // Find a menu by its ID
+    @SuppressWarnings("null")
     Optional<Menu> findById(Long menuId);
 
     // Find a menu by its name
@@ -20,5 +22,5 @@ public interface MenuRepository extends JpaRepository<Menu, Long> {
     List<Menu> findByRestaurantId(Long restaurantId);
 
     // Delete a menu by its ID
-    void deleteById(Long menuId);
+    void deleteById(@SuppressWarnings("null") Long menuId);
 }
